--- a/hw10_program_optimization/stats_test.go
+++ b/hw10_program_optimization/stats_test.go
@@ -13,7 +13,10 @@ var data string = `{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":
 {"Id":2,"Name":"Jesse Vasquez","Username":"qRichardson","Email":"mLynch@broWsecat.com","Phone":"9-373-949-64-00","Password":"SiZLeNSGn","Address":"Fulton Hill 80"}
 {"Id":3,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@Browsecat.com","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
 {"Id":4,"Name":"Gregory Reid","Username":"tButler","Email":"5Moore@Teklist.net","Phone":"520-04-16","Password":"r639qLNu","Address":"Sunfield Park 20"}
-{"Id":5,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@Linktype.com","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}`
+{"Id":5,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@Linktype.com","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}
+{"Id":7,"Name":"Janice Zoo","Username":"Test-01","Email":"linktype.com@otus.ru","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}
+{"Id":8,"Name":"Janice Zoo","Username":"Test-02","Email":"linktype.com@otus.com.ru","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}
+`
 
 func TestGetDomainStat(t *testing.T) {
 	t.Run("find 'com'", func(t *testing.T) {